refactored to move UserDto and LoginUserDto into User folder
