+++ Bibliographia.Web.API/Bibliographia.Web.API.csproj
@@ -32,7 +32,6 @@
     <Folder Include="Database\" />
     <Folder Include="Logs\" />
     <Folder Include="Migrations\" />
-    <Folder Include="Models\DataTransfer\User\" />
   </ItemGroup>
 
   <ItemGroup>
