+++ Bibliographia.Web.API/HelperReferences/Online references for help.txt
@@ -16,6 +16,6 @@ nswag openapi2csclient /input:MyWebService.json
 
                        using nswag command line tool
 
-                       nswag openapi2csclient /input:Bibliographia.nswag.json /classname:BibliographiaClient /namespace:Bibliographia.BlazorUI.Server.WebServicesProxy.Base /output:BibliographiaClientProxy.cs
+nswag openapi2csclient /input:Bibliographia.nswag.json /classname:BibliographiaClient /namespace:Bibliographia.BlazorUI.Server.WebServicesProxy.Base /output:BibliographiaClientProxy.cs
 
 nswag openapi2csclient /input:Bibliographia.nswag.json /GenerateClientInterfaces:true /UseBaseUrl:false /classname:BibliographiaClient /namespace:Bibliographia.BlazorUI.Server.WebServicesProxy.Base /output:BibliographiaClient.cs
